requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
dropbox==9.5.0
fire==0.3.1
idna==2.9
python-slugify==4.0.0
requests==2.23.0
six==1.14.0
termcolor==1.1.0
text-unidecode==1.3
tqdm==4.45.0
Unidecode==1.1.1
urllib3==1.25.8
